kamp-core/src/main/kotlin/ch/leadrian/samp/kamp/core/runtime/callback/PlayerTextDrawCallbackListener.kt
package ch.leadrian.samp.kamp.core.runtime.callback

import ch.leadrian.samp.kamp.core.api.callback.CallbackListenerManager
import ch.leadrian.samp.kamp.core.api.callback.OnPlayerClickPlayerTextDrawListener
import ch.leadrian.samp.kamp.core.api.entity.PlayerTextDraw
import javax.annotation.PostConstruct
import javax.inject.Inject
import javax.inject.Singleton

@Singleton
internal class PlayerTextDrawCallbackListener
@Inject
constructor(
        private val callbackListenerManager: CallbackListenerManager
) : OnPlayerClickPlayerTextDrawListener {

    @PostConstruct
    fun initialize() {
        callbackListenerManager.register(this)
    }

    override fun onPlayerClickPlayerTextDraw(textDraw: PlayerTextDraw): OnPlayerClickPlayerTextDrawListener.Result =
            textDraw.onClick()
}